% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ams_iiie_ontology.R
\name{read_ams_iiie_waste_type_shapes}
\alias{read_ams_iiie_waste_type_shapes}
\title{Load AMS-III.E waste type applicability shape}
\usage{
read_ams_iiie_waste_type_shapes()
}
\value{
A \code{sh_shape_graph} as returned by \code{\link[shapeR:read_shacl]{shapeR::read_shacl()}}.
}
\description{
Load AMS-III.E waste type applicability shape
}
